# internal/app/controller/http/v1/openapi.yml
openapi: 3.0.0
info:
  description: "This is a repository to create, view, edit, and share product and materials certificates"
  version: 0.0.3
  title: Online Certificate Repository
  termsOfService: "temp"
  contact:
    email: "s23-cs-srp-gp12@iastate.edu"
#   license:
#     name: Apache 2.0
#     url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
  /user:
    get:
      tags:
        - user
      summary: Find user by *
      description: Returns a single user
      operationId: getUserBy*
      parameters:
        - name: userEmail
          in: query
          description: email of user to return
          required: false
          schema:
            type: string
        - name: userUUID
          in: query
          description: uuid of user to return
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: '#/components/responses/DefaultErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'    
  /role:
    post:
      tags:
        - userRole
      summary: Create a new User Role
      description: "Creates a new user role"
      operationId: addUserRole
      requestBody:
        description: UserRole to add
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUserRole"
      responses:
        "201":
          description: User Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRole"
        "400":
          $ref: '#/components/responses/DefaultErrorResponse'
        "401":
          $ref: '#/components/responses/DefaultUnauthenticatedErrorResponse'
        "403":
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
    get:
      tags:
        - user
      summary: Find roles by *
      description: Returns a the roles associated with the given UUID
      operationId: getRolesBy*
      parameters:
        - name: userUUID
          in: query
          description: UUID of user for which to get roles
          required: false
          schema:
            type: string
        - name: companyUUID
          in: query
          description: uuid of company for which to get
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRole"
        "404":
          $ref: '#/components/responses/DefaultErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'  
  /company:
    post:
      tags:
        - company
      summary: Create a new Company
      description: "Creates a new company"
      operationId: addCompany
      requestBody:
        description: Company to add
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCompany"              
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          $ref: '#/components/responses/DefaultErrorResponse'
        "401":
          $ref: '#/components/responses/DefaultUnauthenticatedErrorResponse'
        "403":
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
    put:
      tags:
        - company
      summary: Update an existing company
      description: "Update an existing company"
      operationId: updateCompany
      requestBody:
        description: Company to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Company"
      responses:
        "200":
          description: Successively updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          $ref: '#/components/responses/DefaultErrorResponse'
        "401":
          $ref: '#/components/responses/DefaultUnauthenticatedErrorResponse'
        "403":
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        "404":
          $ref: '#/components/responses/DefaultErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
  /company/{companyUUID}:
    get:
      tags:
        - company
      summary: Find company by UUID
      description: Returns a single company
      operationId: getCompanyByUUID
      parameters:
        - name: companyUUID
          in: path
          description: UUID of company to return
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          $ref: '#/components/responses/DefaultErrorResponse'
        "401":
          $ref: '#/components/responses/DefaultUnauthenticatedErrorResponse'
        "403":
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        "404":
          $ref: '#/components/responses/DefaultErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
  /itembatch:
    get:
      tags:
        - itembatch
      summary: Find itembatch by *
      description: Returns a the itembatch associated with the given *
      operationId: getItemBatchBy*
      parameters: # We can always add more optional prams to search by 
        - name: itemBatchUUID
          in: query
          description: UUID of the item batch
          required: false
          schema:
            type: string
        - name: companyUUID
          in: query
          description: uuid of company for which to get
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRole"
        "404":
          $ref: '#/components/responses/DefaultErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'  
    post:
      tags:
        - itembatch
      summary: Create a new item batch
      description: "Creates a new item batch"
      operationId: addItemBatch
      requestBody:
        description: item batch to add
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewItemBatch"              
      responses:
        "201":
          description: ItemBatch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemBatch"
        "400":
          $ref: '#/components/responses/DefaultErrorResponse'
        "401":
          $ref: '#/components/responses/DefaultUnauthenticatedErrorResponse'
        "403":
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
    put:
      tags:
        - itembatch
      summary: Update an existing item batch
      description: "Update an existing item batch"
      operationId: updateItemBatch
      requestBody:
        description: item batch to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemBatch"
      responses:
        "200":
          description: Successively updated ItemBatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemBatch"
        "400":
          $ref: '#/components/responses/DefaultErrorResponse'
        "401":
          $ref: '#/components/responses/DefaultUnauthenticatedErrorResponse'
        "403":
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        "404":
          $ref: '#/components/responses/DefaultErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
  /certification:
    get:
      tags:
        - certification
      summary: Find certification by *
      description: Returns a the certification associated with the given *
      operationId: getCertificationBy*
      parameters: # We can always add more optional prams to search by 
        - name: certificationUUID
          in: query
          description: UUID of the item batch
          required: false
          schema:
            type: string
        - name: companyUUID
          in: query
          description: uuid of company for which to get
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRole"
        "404":
          $ref: '#/components/responses/DefaultErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'  
    post:
      tags:
        - certification
      summary: Create a new item batch
      description: "Creates a new item batch"
      operationId: addCertification
      requestBody:
        description: item batch to add
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCertification"              
      responses:
        "201":
          description: Certification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certification"
        "400":
          $ref: '#/components/responses/DefaultErrorResponse'
        "401":
          $ref: '#/components/responses/DefaultUnauthenticatedErrorResponse'
        "403":
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
    put:
      tags:
        - certification
      summary: Update an existing item batch
      description: "Update an existing item batch"
      operationId: updateCertification
      requestBody:
        description: item batch to update
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Certification"
      responses:
        "200":
          description: Successively updated Certification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certification"
        "400":
          $ref: '#/components/responses/DefaultErrorResponse'
        "401":
          $ref: '#/components/responses/DefaultUnauthenticatedErrorResponse'
        "403":
          $ref: '#/components/responses/DefaultUnauthorizedErrorResponse'
        "404":
          $ref: '#/components/responses/DefaultErrorResponse'
        default:
          $ref: '#/components/responses/DefaultErrorResponse'
# externalDocs:
#   description: Find out more about Swagger
#   url: http://swagger.io
servers:
  - url: localhost:8082/api/v1
components:
  responses:
    DefaultErrorResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"      
      description: unexpected error
    DefaultUnauthenticatedErrorResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"      
      description: unauthenticated error
    DefaultUnauthorizedErrorResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"      
      description: unauthorized error
  schemas:
    User:
      type: object
      required:
        - "uuid"
        - "email"
        - "created"
      properties:
        uuid: 
          type: string
          description: uuid of the user
          pattern: ^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$
        email:
          type: string
          description: email of the user
        created:
          type: string
          description: time user was created at
    NewUserRole:
      type: object
      required:
        - companyUUID
        - roleType
      properties:
        companyUUID:
          type: string
        userUUID:
          type: string
        roleType:
          type: string
        approved:
          type: boolean
      additionalProperties: false
    UserRole:
      allOf:
        - $ref: '#/components/schemas/NewUserRole'
        - type: object
          required:
            - "uuid"
          properties:
            uuid: 
              type: string
          additionalProperties: false
      additionalProperties: false
    NewCompany:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the company
          maxLength: 10
          pattern: "[A-Za-z0-9_]"
      additionalProperties: false
    Company:
      allOf:
        - $ref: '#/components/schemas/NewCompany'
        - type: object
          required:
            - "uuid"
          properties:
            uuid: 
              type: string
              description: uuid of the company
              pattern: ^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$
          additionalProperties: false
      additionalProperties: false
    NewItemBatch:
      type: object
      required:
        - itemNumber
        - "company_uuid"
      properties:
        itemNumber:
          type: string
          description: item number of the batch
          pattern: "[A-Za-z0-9_]"
        description:
          type: string
          description: description of item batch
          pattern: "[A-Za-z0-9_]"
        company_uuid: 
          type: string
          description: uuid of the company
          pattern: ^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$
      additionalProperties: false
    ItemBatch:
      allOf:
        - $ref: '#/components/schemas/NewItemBatch'
        - type: object
          required:
            - "uuid"
          properties:
            uuid: 
              type: string
              description: uuid of the item batch
              pattern: ^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$
          additionalProperties: false
      additionalProperties: false
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code 
        message:
          type: string
          description: Error message
          pattern: "[A-Za-z0-9_]"
          maxLength: 128
      additionalProperties: false